requirements.txt
# pip-compile requirements.in
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
contourpy==1.3.2
    # via matplotlib
cycler==0.12.1
    # via matplotlib
fonttools==4.61.1
    # via matplotlib
kiwisolver==1.4.9
    # via matplotlib
matplotlib==3.10.8
    # via
    #   -r requirements.in
    #   seaborn
numpy==2.2.6
    # via
    #   -r requirements.in
    #   contourpy
    #   matplotlib
    #   pandas
    #   seaborn
packaging==25.0
    # via matplotlib
pandas==2.3.3
    # via
    #   -r requirements.in
    #   seaborn
pillow==12.0.0
    # via matplotlib
psycopg2-binary==2.9.11
    # via -r requirements.in
pyparsing==3.2.5
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   -r requirements.in
    #   matplotlib
    #   pandas
pytz==2025.2
    # via pandas
rapidfuzz==3.14.3
    # via -r requirements.in
seaborn==0.13.2
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
tzdata==2025.3
    # via pandas
unidecode==1.4.0
    # via -r requirements.in
